(PC-43544)[API] refactor: move booking export to its own module
