+++ api/src/pcapi/core/bookings/repository.py
@@ -1,24 +1,15 @@
-import csv
 import typing
 from collections.abc import Sequence
 from datetime import date
 from datetime import datetime
 from datetime import time
 from datetime import timedelta
-from decimal import Decimal
-from io import BytesIO
-from io import StringIO
 
 import sqlalchemy as sa
 import sqlalchemy.orm as sa_orm
-import xlsxwriter
-from xlsxwriter.format import Format
-from xlsxwriter.worksheet import Worksheet
 
 from pcapi.core.bookings import constants
 from pcapi.core.bookings import models
-from pcapi.core.bookings import schemas
-from pcapi.core.bookings import utils
 from pcapi.core.categories import subcategories
 from pcapi.core.finance.models import BookingFinanceIncident
 from pcapi.core.finance.models import FinanceIncident
@@ -30,21 +21,8 @@ from pcapi.core.providers.models import VenueProvider
 from pcapi.core.users.models import User
 from pcapi.models import db
 from pcapi.utils import date as date_utils
-from pcapi.utils import export as utils_export
 from pcapi.utils.token import random_token
 
-
-DUO_QUANTITY = 2
-
-
-BOOKING_STATUS_LABELS = {
-    models.BookingStatus.CONFIRMED: "réservé",
-    models.BookingStatus.CANCELLED: "annulé",
-    models.BookingStatus.USED: "validé",
-    models.BookingStatus.PENDING_REIMBURSEMENT: "en cours de remboursement",
-    models.BookingStatus.REIMBURSED: "remboursé",
-    "confirmed": "confirmé",
-}
 
 BOOKING_DATE_STATUS_MAPPING: dict[models.BookingStatusFilter, sa_orm.InstrumentedAttribute] = {
     models.BookingStatusFilter.BOOKED: models.Booking.dateCreated,
@@ -52,33 +30,14 @@ BOOKING_DATE_STATUS_MAPPING: dict[models.BookingStatusFilter, sa_orm.Instrumente
     models.BookingStatusFilter.REIMBURSED: models.Booking.reimbursementDate,
 }
 
-BOOKING_EXPORT_HEADER = [
-    "Structure",
-    "Nom de l’offre",
-    "Localisation",
-    "Date de l'évènement",
-    "EAN",
-    "Prénom du bénéficiaire",
-    "Nom du bénéficiaire",
-    "Email du bénéficiaire",
-    "Téléphone du bénéficiaire",
-    "Date et heure de réservation",
-    "Date et heure de validation",
-    "Contremarque",
-    "Intitulé du tarif",
-    "Prix de la réservation",
-    "Statut de la contremarque",
-    "Date et heure de remboursement",
-    "Type d'offre",
-    "Code postal du bénéficiaire",
-    "Duo",
-]
-
 BOOKING_LOAD_OPTIONS = Sequence[typing.Literal["offerer", "venue", "offer", "address"]]
 
 
-def booking_export_header() -> list[str]:
-    return BOOKING_EXPORT_HEADER
+def duplicate_booking_when_quantity_is_two(bookings_recap_query: sa_orm.Query) -> sa_orm.Query:
+    duplicated_booking_rows = (
+        sa.func.generate_series(1, models.Booking.quantity).table_valued("duplicate_index").lateral()
+    )
+    return bookings_recap_query.join(duplicated_booking_rows, sa.true())
 
 
 def find_by_venue(
@@ -112,7 +71,7 @@ def find_by_venue(
         offer_id=offer_id,
         offerer_address_id=offerer_address_id,
     )
-    bookings_query = _duplicate_booking_when_quantity_is_two(bookings_query)
+    bookings_query = duplicate_booking_when_quantity_is_two(bookings_query)
     bookings_query = (
         bookings_query.order_by(sa.text('"bookedAt" DESC')).offset((page - 1) * per_page_limit).limit(per_page_limit)
     )
@@ -293,7 +252,7 @@ def get_bookings_from_deposit(deposit_id: int) -> Sequence[models.Booking]:
     return db.session.scalars(query).unique().all()
 
 
-def _create_export_query(offer_id: int, event_beginning_date: date) -> sa_orm.Query:
+def export_query(offer_id: int, event_beginning_date: date) -> sa_orm.Query:
     VenueOffererAddress = sa_orm.aliased(offerers_models.OffererAddress)
     VenueAddress = sa_orm.aliased(Address)
 
@@ -361,64 +320,20 @@ def _create_export_query(offer_id: int, event_beginning_date: date) -> sa_orm.Qu
     return query.distinct(models.Booking.id)
 
 
-def export_validated_bookings_by_offer_id(
-    offer_id: int, event_beginning_date: date, export_type: models.BookingExportType
-) -> str | bytes:
-    offer_validated_bookings_query = _create_export_query(offer_id, event_beginning_date)
-    offer_validated_bookings_query = offer_validated_bookings_query.filter(
+def validated_bookings_by_offer_id_query(offer_id: int, event_beginning_date: date) -> sa_orm.Query[models.Booking]:
+    offer_validated_bookings_query = export_query(offer_id, event_beginning_date)
+    return offer_validated_bookings_query.filter(
         sa.or_(
             sa.and_(models.Booking.isConfirmed.is_(True), models.Booking.status != models.BookingStatus.CANCELLED),
             models.Booking.status == models.BookingStatus.USED,
         )
     )
-    if export_type == models.BookingExportType.EXCEL:
-        return _write_bookings_to_excel(offer_validated_bookings_query)
-    return _write_bookings_to_csv(offer_validated_bookings_query)
-
-
-def export_bookings_by_offer_id(
-    offer_id: int, event_beginning_date: date, export_type: models.BookingExportType
-) -> str | bytes:
-    offer_bookings_query = _create_export_query(offer_id, event_beginning_date)
-    if export_type == models.BookingExportType.EXCEL:
-        return _write_bookings_to_excel(offer_bookings_query)
-    return _write_bookings_to_csv(offer_bookings_query)
-
-
-def get_export(
-    *,
-    pro_user_id: int,
-    venue_ids: list[int],
-    booking_period: tuple[date, date] | None = None,
-    status_filter: models.BookingStatusFilter | None = models.BookingStatusFilter.BOOKED,
-    event_date: date | None = None,
-    offer_id: int | None = None,
-    offerer_address_id: int | None = None,
-    export_type: models.BookingExportType | None = models.BookingExportType.CSV,
-) -> str | bytes:
-    bookings_query = _get_filtered_booking_report(
-        pro_user_id=pro_user_id,
-        venue_ids=venue_ids,
-        period=booking_period,
-        status_filter=status_filter,
-        event_date=event_date,
-        offer_id=offer_id,
-        offerer_address_id=offerer_address_id,
-    )
-    bookings_query = _duplicate_booking_when_quantity_is_two(bookings_query)
-    if export_type == models.BookingExportType.EXCEL:
-        return _serialize_excel_report(bookings_query)
-    return _serialize_csv_report(bookings_query)
 
 
 def field_to_venue_timezone(
     field: sa_orm.InstrumentedAttribute, column: sa_orm.Mapped[typing.Any] | sa.sql.functions.Function
 ) -> sa.Cast[date]:
     return sa.cast(sa.func.timezone(column, sa.func.timezone("UTC", field)), sa.Date)
-
-
-def serialize_offer_type_educational_or_individual(offer_is_educational: bool) -> str:
-    return "offre collective" if offer_is_educational else "offre grand public"
 
 
 def _get_filtered_bookings_query(
@@ -521,7 +436,7 @@ def _get_filtered_bookings_count(
     return bookings_count.scalar()
 
 
-def _get_filtered_booking_report(
+def get_filtered_booking_report(
     *,
     pro_user_id: int,
     venue_ids: list[int],
@@ -655,258 +570,6 @@ def _get_filtered_booking_pro(
     return bookings_query
 
 
-def _duplicate_booking_when_quantity_is_two(bookings_recap_query: sa_orm.Query) -> sa_orm.Query:
-    duplicated_booking_rows = (
-        sa.func.generate_series(1, models.Booking.quantity).table_valued("duplicate_index").lateral()
-    )
-    return bookings_recap_query.join(duplicated_booking_rows, sa.true())
-
-
-def _get_booking_status(status: models.BookingStatus, is_confirmed: bool) -> str:
-    cancellation_limit_date_exists_and_past = is_confirmed
-    if cancellation_limit_date_exists_and_past and status == models.BookingStatus.CONFIRMED:
-        return BOOKING_STATUS_LABELS["confirmed"]
-    return BOOKING_STATUS_LABELS[status]
-
-
-def get_booking_price(booking: schemas.ExportBookingsQueryResult) -> Decimal:
-    """
-    Retourne le prix de la réservation, converti en CFP si le bénéficiaire est calédonien.
-    """
-    if hasattr(booking, "is_caledonian") and booking.is_caledonian:
-        return utils.convert_euro_to_pacific_franc(booking.amount)
-    return booking.amount
-
-
-def _write_bookings_to_csv(query: sa_orm.Query) -> str:
-    output = StringIO()
-    writer = csv.writer(output, dialect=csv.excel, delimiter=";", quoting=csv.QUOTE_NONNUMERIC)
-    writer.writerow(booking_export_header())
-    for booking in query.yield_per(1000):
-        booking = typing.cast(schemas.ExportBookingsQueryResult, booking)
-
-        if booking.quantity == DUO_QUANTITY:
-            _write_csv_row(writer, booking, "DUO 1")
-            _write_csv_row(writer, booking, "DUO 2")
-        else:
-            _write_csv_row(writer, booking, "Non")
-
-    return output.getvalue()
-
-
-def _write_csv_row(csv_writer: typing.Any, booking: schemas.ExportBookingsQueryResult, booking_duo_column: str) -> None:
-    booking_price = get_booking_price(booking)
-    row: tuple[typing.Any, ...] = (
-        booking.venueName,
-        booking.offerName,
-        f"{booking.locationName} - {booking.locationStreet} {booking.locationPostalCode} {booking.locationCity}",
-        utils.convert_booking_dates_utc_to_venue_timezone(booking.stockBeginningDatetime, booking),
-        booking.ean,
-        booking.beneficiaryFirstName,
-        booking.beneficiaryLastName,
-        booking.beneficiaryEmail,
-        booking.beneficiaryPhoneNumber,
-        utils.convert_booking_dates_utc_to_venue_timezone(booking.bookedAt, booking),
-        utils.convert_booking_dates_utc_to_venue_timezone(booking.usedAt, booking),
-        get_booking_token(
-            booking.token,
-            booking.status,
-            booking.isExternal,
-            booking.stockBeginningDatetime,
-        ),
-        booking.priceCategoryLabel or "",
-        booking_price,
-        _get_booking_status(booking.status, booking.isConfirmed),
-        utils.convert_booking_dates_utc_to_venue_timezone(booking.reimbursedAt, booking),
-        serialize_offer_type_educational_or_individual(offer_is_educational=False),
-        booking.beneficiaryPostalCode or "",
-        booking_duo_column,
-    )
-    csv_writer.writerow(row)
-
-
-def _write_bookings_to_excel(query: sa_orm.Query) -> bytes:
-    output = BytesIO()
-    workbook = xlsxwriter.Workbook(output)
-
-    bold = workbook.add_format(utils_export.EXCEL_BOLD_FORMAT)
-    currency_format_eur = workbook.add_format(utils_export.EXCEL_CURRENCY_FORMAT)
-    currency_format_cfp = workbook.add_format(utils_export.EXCEL_CFP_FORMAT)
-    col_width = utils_export.EXCEL_COL_WIDTH
-
-    worksheet = workbook.add_worksheet()
-    row = 0
-
-    for col_num, title in enumerate(booking_export_header()):
-        worksheet.write(row, col_num, title, bold)
-        worksheet.set_column(col_num, col_num, col_width)
-
-    row = 1
-    for booking in query.yield_per(1000):
-        booking = typing.cast(schemas.ExportBookingsQueryResult, booking)
-
-        if booking.quantity == DUO_QUANTITY:
-            _write_excel_row(
-                worksheet,
-                row,
-                booking,
-                currency_format_cfp if getattr(booking, "is_caledonian", False) else currency_format_eur,
-                "DUO 1",
-            )
-            row += 1
-            _write_excel_row(
-                worksheet,
-                row,
-                booking,
-                currency_format_cfp if getattr(booking, "is_caledonian", False) else currency_format_eur,
-                "DUO 2",
-            )
-        else:
-            _write_excel_row(
-                worksheet,
-                row,
-                booking,
-                currency_format_cfp if getattr(booking, "is_caledonian", False) else currency_format_eur,
-                "Non",
-            )
-        row += 1
-    workbook.close()
-    return output.getvalue()
-
-
-def _write_excel_row(
-    worksheet: Worksheet, row: int, booking: schemas.ExportBookingsQueryResult, currency_format: Format, duo_column: str
-) -> None:
-    booking_price = get_booking_price(booking)
-    worksheet.write(row, 0, booking.venueName)
-    worksheet.write(row, 1, booking.offerName)
-    worksheet.write(
-        row, 2, str(utils.convert_booking_dates_utc_to_venue_timezone(booking.stockBeginningDatetime, booking))
-    )
-    worksheet.write(row, 3, booking.ean)
-    worksheet.write(row, 4, booking.beneficiaryFirstName)
-    worksheet.write(row, 5, booking.beneficiaryLastName)
-    worksheet.write(row, 6, booking.beneficiaryEmail)
-    worksheet.write(row, 7, booking.beneficiaryPhoneNumber)
-    worksheet.write(row, 8, str(utils.convert_booking_dates_utc_to_venue_timezone(booking.bookedAt, booking)))
-    worksheet.write(row, 9, str(utils.convert_booking_dates_utc_to_venue_timezone(booking.usedAt, booking)))
-    worksheet.write(
-        row,
-        10,
-        get_booking_token(
-            booking.token,
-            booking.status,
-            booking.isExternal,
-            booking.stockBeginningDatetime,
-        ),
-    )
-    worksheet.write(row, 11, booking.priceCategoryLabel)
-    worksheet.write(row, 12, booking_price, currency_format)
-    worksheet.write(row, 13, _get_booking_status(booking.status, booking.isConfirmed))
-    worksheet.write(row, 14, str(utils.convert_booking_dates_utc_to_venue_timezone(booking.reimbursedAt, booking)))
-    worksheet.write(row, 15, serialize_offer_type_educational_or_individual(offer_is_educational=False))
-    worksheet.write(row, 16, booking.beneficiaryPostalCode)
-    worksheet.write(
-        row,
-        17,
-        duo_column,
-    )
-
-
-def _serialize_csv_report(query: sa_orm.Query) -> str:
-    output = StringIO()
-    writer = csv.writer(output, dialect=csv.excel, delimiter=";", quoting=csv.QUOTE_NONNUMERIC)
-    writer.writerow(booking_export_header())
-    for booking in query.yield_per(1000):
-        booking = typing.cast(schemas.ExportBookingsQueryResult, booking)
-
-        booking_price = get_booking_price(booking)
-        row: tuple[typing.Any, ...] = (
-            booking.venueName,
-            booking.offerName,
-            f"{booking.locationName} - {booking.locationStreet} {booking.locationPostalCode} {booking.locationCity}",
-            utils.convert_booking_dates_utc_to_venue_timezone(booking.stockBeginningDatetime, booking),
-            booking.ean,
-            booking.beneficiaryFirstName,
-            booking.beneficiaryLastName,
-            booking.beneficiaryEmail,
-            booking.beneficiaryPhoneNumber,
-            utils.convert_booking_dates_utc_to_venue_timezone(booking.bookedAt, booking),
-            utils.convert_booking_dates_utc_to_venue_timezone(booking.usedAt, booking),
-            get_booking_token(
-                booking.token,
-                booking.status,
-                booking.isExternal,
-                booking.stockBeginningDatetime,
-            ),
-            booking.priceCategoryLabel or "",
-            booking_price,
-            _get_booking_status(booking.status, booking.isConfirmed),
-            utils.convert_booking_dates_utc_to_venue_timezone(booking.reimbursedAt, booking),
-            # This method is still used in the old Payment model
-            serialize_offer_type_educational_or_individual(offer_is_educational=False),
-            booking.beneficiaryPostalCode or "",
-            "Oui" if booking.quantity == DUO_QUANTITY else "Non",
-        )
-        writer.writerow(row)
-
-    return output.getvalue()
-
-
-def _serialize_excel_report(query: sa_orm.Query) -> bytes:
-    output = BytesIO()
-    workbook = xlsxwriter.Workbook(output)
-
-    bold = workbook.add_format(utils_export.EXCEL_BOLD_FORMAT)
-    currency_format_eur = workbook.add_format(utils_export.EXCEL_CURRENCY_FORMAT)
-    currency_format_cfp = workbook.add_format(utils_export.EXCEL_CFP_FORMAT)
-    col_width = utils_export.EXCEL_COL_WIDTH
-
-    worksheet = workbook.add_worksheet()
-    row = 0
-
-    for col_num, title in enumerate(booking_export_header()):
-        worksheet.write(row, col_num, title, bold)
-        worksheet.set_column(col_num, col_num, col_width)
-    row = 1
-    data: tuple[typing.Any, ...]
-    for booking in query.yield_per(1000):
-        booking = typing.cast(schemas.ExportBookingsQueryResult, booking)
-
-        booking_price = get_booking_price(booking)
-        if hasattr(booking, "is_caledonian") and booking.is_caledonian:
-            currency_format = currency_format_cfp
-        else:
-            currency_format = currency_format_eur
-        data = (
-            booking.venueName,
-            booking.offerName,
-            f"{booking.locationName} - {booking.locationStreet} {booking.locationPostalCode} {booking.locationCity}",
-            str(utils.convert_booking_dates_utc_to_venue_timezone(booking.stockBeginningDatetime, booking)),
-            booking.ean,
-            booking.beneficiaryFirstName,
-            booking.beneficiaryLastName,
-            booking.beneficiaryEmail,
-            booking.beneficiaryPhoneNumber,
-            str(utils.convert_booking_dates_utc_to_venue_timezone(booking.bookedAt, booking)),
-            str(utils.convert_booking_dates_utc_to_venue_timezone(booking.usedAt, booking)),
-            get_booking_token(booking.token, booking.status, booking.isExternal, booking.stockBeginningDatetime),
-            booking.priceCategoryLabel,
-            booking_price,
-            _get_booking_status(booking.status, booking.isConfirmed),
-            str(utils.convert_booking_dates_utc_to_venue_timezone(booking.reimbursedAt, booking)),
-            serialize_offer_type_educational_or_individual(offer_is_educational=False),
-            booking.beneficiaryPostalCode,
-            "Oui" if booking.quantity == DUO_QUANTITY else "Non",
-        )
-        worksheet.write_row(row, 0, data)
-        worksheet.set_column(13, 13, cell_format=currency_format)
-        row += 1
-
-    workbook.close()
-    return output.getvalue()
-
-
 def get_soon_expiring_bookings(expiration_days_delta: int) -> typing.Generator[models.Booking]:
     """Find bookings expiring in exactly `expiration_days_delta` days"""
     query = (
@@ -1010,24 +673,3 @@ def get_external_bookings_by_cinema_id_and_barcodes(
         .filter(models.ExternalBooking.barcode.in_(barcodes))
         .all()
     )
-
-
-def get_booking_token(
-    booking_token: str,
-    booking_status: models.BookingStatus,
-    booking_is_external: bool,
-    event_beginning_datetime: datetime | None,
-) -> str | None:
-    if (
-        not event_beginning_datetime
-        and booking_status
-        not in [
-            models.BookingStatus.PENDING_REIMBURSEMENT,
-            models.BookingStatus.REIMBURSED,
-            models.BookingStatus.CANCELLED,
-            models.BookingStatus.USED,
-        ]
-        or booking_is_external
-    ):
-        return None
-    return booking_token
